requirements.txt
google-api-python-client==2.118.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
requests==2.31.0
isodate==0.6.1
python-dotenv==1.0.1
